+++ sso/awssso.go
@@ -20,18 +20,19 @@ package sso
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"math/rand"
 	"reflect"
 	"strconv"
 	"sync"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/aws/retry"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/service/sso"
-	"github.com/aws/aws-sdk-go-v2/service/sso/types"
+	ssotypes "github.com/aws/aws-sdk-go-v2/service/sso/types"
 	"github.com/aws/aws-sdk-go-v2/service/ssooidc"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
 	"github.com/davecgh/go-spew/spew"
@@ -41,9 +42,8 @@ import (
 	"github.com/synfinatic/gotable"
 )
 
-const (
-	MAX_RETRY_DELAY = 10.0
-)
+var MAX_RETRY_ATTEMPTS int = 10
+var MAX_BACKOFF_SECONDS int = 5
 
 // Necessary for mocking
 type SsoOidcAPI interface {
@@ -82,14 +82,29 @@ type AWSSSO struct {
 }
 
 func NewAWSSSO(s *SSOConfig, store *storage.SecureStorage) *AWSSSO {
+	var maxRetry int = MAX_RETRY_ATTEMPTS
+	if s.MaxRetry > 0 {
+		maxRetry = s.MaxRetry
+	}
+	var maxBackoff int = MAX_BACKOFF_SECONDS
+	if s.MaxBackoff > 0 {
+		maxBackoff = s.MaxBackoff
+	}
+	log.Debugf("loading SSO using %d retries and max %dsec backoff", maxRetry, maxBackoff)
+
+	r := retry.NewStandard(func(o *retry.StandardOptions) {
+		o.MaxAttempts = maxRetry
+		o.MaxBackoff = time.Duration(maxBackoff) * time.Second
+	})
+
 	oidcSession := ssooidc.New(ssooidc.Options{
-		Region:           s.SSORegion,
-		RetryMaxAttempts: 5,
+		Region:  s.SSORegion,
+		Retryer: r,
 	})
 
 	ssoSession := sso.New(sso.Options{
-		Region:           s.SSORegion,
-		RetryMaxAttempts: 5,
+		Region:  s.SSORegion,
+		Retryer: r,
 	})
 
 	as := AWSSSO{
@@ -148,7 +163,7 @@ func (ri RoleInfo) GetAccountId64() int64 {
 
 // GetRoles fetches all the AWS SSO IAM Roles for the given AWS Account
 // Code is running up to X Threads via cache.processSSORoles()
-// and we must stricly protect our as.Roles[] dict
+// and we must stricly protect reads & writes to our as.Roles[] dict
 func (as *AWSSSO) GetRoles(account AccountInfo) ([]RoleInfo, error) {
 	as.rolesLock.RLock()
 	roles, ok := as.Roles[account.AccountId]
@@ -161,30 +176,37 @@ func (as *AWSSSO) GetRoles(account AccountInfo) ([]RoleInfo, error) {
 	as.Roles[account.AccountId] = []RoleInfo{}
 	as.rolesLock.Unlock()
 
-	// why is this input value locked???
-	as.tokenLock.RLock()
+	// must lock this becacase the access token can change
+	as.tokenLock.Lock()
 	input := sso.ListAccountRolesInput{
 		AccessToken: aws.String(as.Token.AccessToken),
 		AccountId:   aws.String(account.AccountId),
 		MaxResults:  aws.Int32(1000),
 	}
-	as.tokenLock.RUnlock()
+	as.tokenLock.Unlock()
 
-	output := as.ListAccountRoles(&input)
+	output, err := as.ListAccountRoles(&input)
+	if err != nil {
+		// failed... give up
+		as.rolesLock.RLock()
+		defer as.rolesLock.RUnlock()
+		return as.Roles[account.AccountId], err
+	}
 
 	// Process the output
 	for i, r := range output.RoleList {
-		if err := as.makeRoleInfo(account, i, r); err != nil {
+		as.makeRoleInfo(account, i, r)
+	}
+
+	for aws.ToString(output.NextToken) != "" {
+		input.NextToken = output.NextToken
+		output, err = as.ListAccountRoles(&input)
+		if err != nil {
 			// failed... give up
 			as.rolesLock.RLock()
 			defer as.rolesLock.RUnlock()
 			return as.Roles[account.AccountId], err
 		}
-	}
-
-	for aws.ToString(output.NextToken) != "" {
-		input.NextToken = output.NextToken
-		output = as.ListAccountRoles(&input)
 
 		as.rolesLock.RLock()
 		roleCount := len(as.Roles[account.AccountId])
@@ -192,11 +214,7 @@ func (as *AWSSSO) GetRoles(account AccountInfo) ([]RoleInfo, error) {
 
 		for i, r := range output.RoleList {
 			x := roleCount + i
-			if err := as.makeRoleInfo(account, x, r); err != nil {
-				as.rolesLock.RLock()
-				defer as.rolesLock.RUnlock()
-				return as.Roles[account.AccountId], err
-			}
+			as.makeRoleInfo(account, x, r)
 		}
 	}
 	as.rolesLock.RLock()
@@ -204,45 +222,81 @@ func (as *AWSSSO) GetRoles(account AccountInfo) ([]RoleInfo, error) {
 	return as.Roles[account.AccountId], nil
 }
 
-// ListAccountRoles is a wrapper around sso.ListAccountRoles which does our retry logic
-func (as *AWSSSO) ListAccountRoles(input *sso.ListAccountRolesInput) *sso.ListAccountRolesOutput {
-	output, err := as.sso.ListAccountRoles(context.TODO(), input)
-	if err != nil {
-		switch err.(type) {
-		case *types.TooManyRequestsException:
-			// 429 errors indicate we should wait and then try again
-			sleepSec := rand.Float32() * MAX_RETRY_DELAY // #nosec sleep random amount up MAX_RETRY_DELAY nosec
-			log.Errorf("TooManyRequestsException failure; trying again in %0.2f seconds...", sleepSec)
-			time.Sleep(time.Duration(sleepSec) * time.Second)
+func (as *AWSSSO) ListAccounts(input *sso.ListAccountsInput) (*sso.ListAccountsOutput, error) {
+	var err error = errors.New("foo")
+	var output *sso.ListAccountsOutput
 
-		case *types.UnauthorizedException:
-			// if we have to re-auth, keep hold everyone else up
-			as.rolesLock.RLock()
-			log.Errorf("AccessToken Unauthorized Error; refreshing: %s", err.Error())
+	for cnt := 0; err != nil && cnt <= MAX_RETRY_ATTEMPTS; cnt++ {
+		output, err = as.sso.ListAccounts(context.TODO(), input)
+		if err != nil {
+			var tmr *ssotypes.TooManyRequestsException
+			var ue *ssotypes.UnauthorizedException
+			switch {
+			case errors.As(err, &ue):
+				// sometimes our AccessToken is invalid so try a new one once?
+				// if we have to re-auth, hold everyone else up since that will reduce other failures
+				as.rolesLock.Lock()
+				log.Errorf("AccessToken Unauthorized Error; refreshing: %s", err.Error())
 
-			if err = as.reauthenticate(); err != nil {
-				// failed hard
-				log.WithError(err).Fatalf("Unexpected auth failure")
+				if err = as.reauthenticate(); err != nil {
+					// fail hard now
+					return output, err
+				}
+				input.AccessToken = aws.String(as.Token.AccessToken)
+				as.rolesLock.Unlock()
+			case errors.As(err, &tmr):
+				// try again
+				log.Warnf("Exceeded MaxRetry/MaxBackoff.  Consider tuning values.")
+				time.Sleep(time.Duration(MAX_BACKOFF_SECONDS) * time.Second)
+
+			default:
+				log.WithError(err).Error("Unexpected error")
 			}
-			as.rolesLock.RUnlock()
-			input.AccessToken = aws.String(as.Token.AccessToken)
-
-		default:
-			log.WithError(err).Fatalf("Unexpected error")
-		}
-
-		// retry the ListAccountRoles after doing our needful...
-		if output, err = as.sso.ListAccountRoles(context.TODO(), input); err != nil {
-			// failed again.  Give up.
-			log.WithError(err).Fatalf("Unexpected error after retry")
 		}
 	}
+	return output, err
+}
 
-	return output
+// ListAccountRoles is a wrapper around sso.ListAccountRoles which does our retry logic
+func (as *AWSSSO) ListAccountRoles(input *sso.ListAccountRolesInput) (*sso.ListAccountRolesOutput, error) {
+	var err error = errors.New("foo")
+	var output *sso.ListAccountRolesOutput
+
+	for cnt := 0; err != nil && cnt <= MAX_RETRY_ATTEMPTS; cnt++ {
+		output, err = as.sso.ListAccountRoles(context.TODO(), input)
+
+		if err != nil {
+			var tmr *ssotypes.TooManyRequestsException
+			var ue *ssotypes.UnauthorizedException
+			switch {
+			case errors.As(err, &ue):
+				// sometimes our AccessToken is invalid so try a new one once?
+				// if we have to re-auth, hold everyone else up since that will reduce other failures
+				as.rolesLock.Lock()
+				log.Errorf("AccessToken Unauthorized Error; refreshing: %s", err.Error())
+
+				if err = as.reauthenticate(); err != nil {
+					// fail hard now
+					log.WithError(err).Fatalf("Unexpected auth failure")
+				}
+				input.AccessToken = aws.String(as.Token.AccessToken)
+				as.rolesLock.Unlock()
+
+			case errors.As(err, &tmr):
+				// try again
+				log.Warnf("Exceeded MaxRetry/MaxBackoff.  Consider tuning values.")
+				time.Sleep(time.Duration(MAX_BACKOFF_SECONDS) * time.Second)
+
+			default:
+				log.WithError(err).Error("Unexpected error")
+			}
+		}
+	}
+	return output, err
 }
 
 // makeRoleInfo takes the sso.types.RoleInfo and adds it onto our as.Roles[accountId] list
-func (as *AWSSSO) makeRoleInfo(account AccountInfo, i int, r types.RoleInfo) error {
+func (as *AWSSSO) makeRoleInfo(account AccountInfo, i int, r ssotypes.RoleInfo) {
 	var via string
 
 	aId, _ := strconv.ParseInt(account.AccountId, 10, 64)
@@ -264,7 +318,6 @@ func (as *AWSSSO) makeRoleInfo(account AccountInfo, i int, r types.RoleInfo) err
 		StartUrl:     as.StartUrl,
 		Via:          via,
 	})
-	return nil
 }
 
 type AccountInfo struct {
@@ -300,17 +353,9 @@ func (as *AWSSSO) GetAccounts() ([]AccountInfo, error) {
 		AccessToken: aws.String(as.Token.AccessToken),
 		MaxResults:  aws.Int32(1000),
 	}
-	output, err := as.sso.ListAccounts(context.TODO(), &input)
+	output, err := as.ListAccounts(&input)
 	if err != nil {
-		// sometimes our AccessToken is invalid so try a new one once?
-		log.Errorf("Unexpected AccessToken failure; refreshing: %s", err)
-		if err = as.reauthenticate(); err != nil {
-			return as.Accounts, err
-		}
-		input.AccessToken = aws.String(as.Token.AccessToken)
-		if output, err = as.sso.ListAccounts(context.TODO(), &input); err != nil {
-			return as.Accounts, err
-		}
+		return as.Accounts, err
 	}
 
 	for i, r := range output.AccountList {
@@ -324,7 +369,7 @@ func (as *AWSSSO) GetAccounts() ([]AccountInfo, error) {
 
 	for aws.ToString(output.NextToken) != "" {
 		input.NextToken = output.NextToken
-		output, err = as.sso.ListAccounts(context.TODO(), &input)
+		output, err = as.ListAccounts(&input)
 		if err != nil {
 			return as.Accounts, err
 		}